requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
pillow==8.4.0
    # via -r requirements.in
